Move ModelSchemas to test_model_schema.py
